+++ SecadorBotas/Frames/FrmConfPeriodoActividad3.cs
@@ -65,14 +65,24 @@ namespace SecadorBotas.Frames
 
                 else
                 {
-                    //LLAMADA DEL METODO INSERTAR ENCENDIDO DE SECADOR 3
-                    conex.InsertarEncendido3(dia, hora, minuto);
 
-                    //LLAMADA A METODO QUE CARGA EL GRID
-                    CargarDataGridON();
+                    DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Guardar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
 
-                    //MENSAJE POR PANTALLA
-                    MessageBox.Show("Encendido automático guardado!");
+                    if (result == DialogResult.Yes)
+                    {
+                        //LLAMADA DEL METODO INSERTAR ENCENDIDO DE SECADOR 3
+                        conex.InsertarEncendido3(dia, hora, minuto);
+
+                        //LLAMADA A METODO QUE CARGA EL GRID
+                        CargarDataGridON();
+
+                        //MENSAJE POR PANTALLA
+                        MessageBox.Show("Encendido automático guardado!");
+                    }
+                    else
+                    {
+
+                    }
                 }
             }
         }
@@ -118,20 +128,30 @@ namespace SecadorBotas.Frames
 
             {
 
-                string query = "delete from EncendidoSecador3 where id = @Id";
+                DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Eliminar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+                if (result == DialogResult.Yes)
+                {
+                    string query = "delete from EncendidoSecador3 where id = @Id";
 
-                SqlCommand cmd = new SqlCommand(query, cnn);
-                cnn.Open();
+                    SqlCommand cmd = new SqlCommand(query, cnn);
+                    cnn.Open();
 
-                cmd.Parameters.AddWithValue("@Id", Id);
-                cmd.ExecuteNonQuery();
+                    cmd.Parameters.AddWithValue("@Id", Id);
+                    cmd.ExecuteNonQuery();
+
+                    dataGridViewON.Rows.Remove(dataGridViewON.CurrentRow);
+
+                    //MENSAJE POR PANTALLA
+                    MessageBox.Show("Encendido automático removido!");
+                }
+                else
+                {
+
+                }
 
             }
 
-            dataGridViewON.Rows.Remove(dataGridViewON.CurrentRow);
-
-            //MENSAJE POR PANTALLA
-            MessageBox.Show("Encendido automático removido!");
+            
         }
 
         private void pbAumentaONhrs_Click(object sender, EventArgs e)
@@ -202,14 +222,23 @@ namespace SecadorBotas.Frames
                 else
                 {
 
-                    //LLAMADA DEL METODO INSERTAR APAGADO DE SECADOR 3
-                    conex.InsertarApagado3(dia, hora, minuto);
+                    DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Guardar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
 
-                    //LLAMADA A METODO QUE CARGA EL GRID
-                    CargarDataGridOFF();
+                    if (result == DialogResult.Yes)
+                    {
+                        //LLAMADA DEL METODO INSERTAR APAGADO DE SECADOR 3
+                        conex.InsertarApagado3(dia, hora, minuto);
 
-                    //MENSAJE POR PANTALLA
-                    MessageBox.Show("Apagado automático guardado!");
+                        //LLAMADA A METODO QUE CARGA EL GRID
+                        CargarDataGridOFF();
+
+                        //MENSAJE POR PANTALLA
+                        MessageBox.Show("Apagado automático guardado!");
+                    }
+                    else
+                    {
+
+                    }
                 }
             }
         }
@@ -228,21 +257,30 @@ namespace SecadorBotas.Frames
             using (SqlConnection cnn = new SqlConnection(sqlConexionWebConfig))
 
             {
+                DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Eliminar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+                if (result == DialogResult.Yes)
+                {
+                    string query = "delete from ApagadoSecador3 where idOff = @Id";
 
-                string query = "delete from ApagadoSecador3 where idOff = @Id";
+                    SqlCommand cmd = new SqlCommand(query, cnn);
+                    cnn.Open();
 
-                SqlCommand cmd = new SqlCommand(query, cnn);
-                cnn.Open();
+                    cmd.Parameters.AddWithValue("@Id", IdOff);
+                    cmd.ExecuteNonQuery();
 
-                cmd.Parameters.AddWithValue("@Id", IdOff);
-                cmd.ExecuteNonQuery();
+                    dataGridViewOFF.Rows.Remove(dataGridViewOFF.CurrentRow);
+
+                    //MENSAJE POR PANTALLA
+                    MessageBox.Show("Apagado automático removido!");
+                }
+                else
+                {
+
+                }
 
             }
 
-            dataGridViewOFF.Rows.Remove(dataGridViewOFF.CurrentRow);
-
-            //MENSAJE POR PANTALLA
-            MessageBox.Show("Apagado automático removido!");
+            
         }
 
         private void pictureBox18_Click(object sender, EventArgs e)
