+++ SecadorBotas/Frames/FrmConfPeriodoActividad5.cs
@@ -67,14 +67,24 @@ namespace SecadorBotas.Frames
 
                 else
                 {
-                    //LLAMADA DEL METODO INSERTAR ENCENDIDO DE SECADOR 6
-                    conex.InsertarEncendido5(dia, hora, minuto);
+                    DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Guardar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
 
-                    //LLAMADA A METODO QUE CARGA EL GRID
-                    CargarDataGridON();
+                    if (result == DialogResult.Yes)
+                    {
+                        //LLAMADA DEL METODO INSERTAR ENCENDIDO DE SECADOR 6
+                        conex.InsertarEncendido5(dia, hora, minuto);
 
-                    //MENSAJE POR PANTALLA
-                    MessageBox.Show("Encendido automático guardado!");
+                        //LLAMADA A METODO QUE CARGA EL GRID
+                        CargarDataGridON();
+
+                        //MENSAJE POR PANTALLA
+                        MessageBox.Show("Encendido automático guardado!");
+
+                    }
+                    else
+                    {
+
+                    }
                 }
             }
         }
@@ -122,21 +132,30 @@ namespace SecadorBotas.Frames
             using (SqlConnection cnn = new SqlConnection(sqlConexionWebConfig))
 
             {
+                DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Eliminar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+                if (result == DialogResult.Yes)
+                {
+                    string query = "delete from EncendidoSecador5 where id = @Id";
 
-                string query = "delete from EncendidoSecador5 where id = @Id";
+                    SqlCommand cmd = new SqlCommand(query, cnn);
+                    cnn.Open();
 
-                SqlCommand cmd = new SqlCommand(query, cnn);
-                cnn.Open();
+                    cmd.Parameters.AddWithValue("@Id", Id);
+                    cmd.ExecuteNonQuery();
 
-                cmd.Parameters.AddWithValue("@Id", Id);
-                cmd.ExecuteNonQuery();
+                    dataGridViewON.Rows.Remove(dataGridViewON.CurrentRow);
+
+                    //MENSAJE POR PANTALLA
+                    MessageBox.Show("Encendido automático removido!");
+                }
+                else
+                {
+
+                }
 
             }
 
-            dataGridViewON.Rows.Remove(dataGridViewON.CurrentRow);
-
-            //MENSAJE POR PANTALLA
-            MessageBox.Show("Encendido automático removido!");
+            
         }
 
         private void pbAumentaONhrs_Click(object sender, EventArgs e)
@@ -207,15 +226,23 @@ namespace SecadorBotas.Frames
                 }
                 else
                 {
+                    DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Guardar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
 
-                    //LLAMADA DEL METODO INSERTAR APAGADO DE SECADOR 7
-                    conex.InsertarApagado5(dia, hora, minuto);
+                    if (result == DialogResult.Yes)
+                    {
+                        //LLAMADA DEL METODO INSERTAR APAGADO DE SECADOR 7
+                        conex.InsertarApagado5(dia, hora, minuto);
 
-                    //LLAMADA A METODO QUE CARGA EL GRID
-                    CargarDataGridOFF();
+                        //LLAMADA A METODO QUE CARGA EL GRID
+                        CargarDataGridOFF();
 
-                    //MENSAJE POR PANTALLA
-                    MessageBox.Show("Apagado automático guardado!");
+                        //MENSAJE POR PANTALLA
+                        MessageBox.Show("Apagado automático guardado!");
+                    }
+                    else
+                    {
+
+                    }
                 }
             }
         }
@@ -235,20 +262,30 @@ namespace SecadorBotas.Frames
 
             {
 
-                string query = "delete from ApagadoSecador5 where idOff = @Id";
+                DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Eliminar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+                if (result == DialogResult.Yes)
+                {
+                    string query = "delete from ApagadoSecador5 where idOff = @Id";
 
-                SqlCommand cmd = new SqlCommand(query, cnn);
-                cnn.Open();
+                    SqlCommand cmd = new SqlCommand(query, cnn);
+                    cnn.Open();
 
-                cmd.Parameters.AddWithValue("@Id", IdOff);
-                cmd.ExecuteNonQuery();
+                    cmd.Parameters.AddWithValue("@Id", IdOff);
+                    cmd.ExecuteNonQuery();
+
+                    dataGridViewOFF.Rows.Remove(dataGridViewOFF.CurrentRow);
+
+                    //MENSAJE POR PANTALLA
+                    MessageBox.Show("Apagado automático removido!");
+                }
+                else
+                {
+
+                }
 
             }
 
-            dataGridViewOFF.Rows.Remove(dataGridViewOFF.CurrentRow);
-
-            //MENSAJE POR PANTALLA
-            MessageBox.Show("Apagado automático removido!");
+            
         }
 
         private void pictureBox18_Click(object sender, EventArgs e)
