+++ SecadorBotas/Frames/FrmConfPeriodoActividad7.cs
@@ -228,21 +228,30 @@ namespace SecadorBotas.Frames
             using (SqlConnection cnn = new SqlConnection(sqlConexionWebConfig))
 
             {
+                DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Eliminar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+                if (result == DialogResult.Yes)
+                {
+                    string query = "delete from EncendidoSecador7 where id = @Id";
 
-                string query = "delete from EncendidoSecador7 where id = @Id";
+                    SqlCommand cmd = new SqlCommand(query, cnn);
+                    cnn.Open();
 
-                SqlCommand cmd = new SqlCommand(query, cnn);
-                cnn.Open();
+                    cmd.Parameters.AddWithValue("@Id", Id);
+                    cmd.ExecuteNonQuery();
 
-                cmd.Parameters.AddWithValue("@Id", Id);
-                cmd.ExecuteNonQuery();
+                    dataGridViewON.Rows.Remove(dataGridViewON.CurrentRow);
+
+                    //MENSAJE POR PANTALLA
+                    MessageBox.Show("Encendido automático removido!");
+                }
+                else
+                {
+
+                }
 
             }
 
-            dataGridViewON.Rows.Remove(dataGridViewON.CurrentRow);
-
-            //MENSAJE POR PANTALLA
-            MessageBox.Show("Encendido automático removido!");
+           
         }
 
         private void btnGuardarON_Click_1(object sender, EventArgs e)
@@ -273,14 +282,24 @@ namespace SecadorBotas.Frames
 
                 else
                 {
-                    //LLAMADA DEL METODO INSERTAR ENCENDIDO DE SECADOR 7
-                    conex.InsertarEncendido7(dia, hora, minuto);
 
-                    //LLAMADA A METODO QUE CARGA EL GRID
-                    CargarDataGridON();
+                    DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Guardar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
 
-                    //MENSAJE POR PANTALLA
-                    MessageBox.Show("Encendido automático guardado!");
+                    if (result == DialogResult.Yes)
+                    {
+                        //LLAMADA DEL METODO INSERTAR ENCENDIDO DE SECADOR 7
+                        conex.InsertarEncendido7(dia, hora, minuto);
+
+                        //LLAMADA A METODO QUE CARGA EL GRID
+                        CargarDataGridON();
+
+                        //MENSAJE POR PANTALLA
+                        MessageBox.Show("Encendido automático guardado!");
+                    }
+                    else
+                    {
+
+                    }
                 }
             }
         }
@@ -300,21 +319,29 @@ namespace SecadorBotas.Frames
             using (SqlConnection cnn = new SqlConnection(sqlConexionWebConfig))
 
             {
+                DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Eliminar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
+                if (result == DialogResult.Yes)
+                {
+                    string query = "delete from ApagadoSecador7 where idOff = @Id";
 
-                string query = "delete from ApagadoSecador7 where idOff = @Id";
+                    SqlCommand cmd = new SqlCommand(query, cnn);
+                    cnn.Open();
 
-                SqlCommand cmd = new SqlCommand(query, cnn);
-                cnn.Open();
+                    cmd.Parameters.AddWithValue("@Id", IdOff);
+                    cmd.ExecuteNonQuery();
+                    dataGridViewOFF.Rows.Remove(dataGridViewOFF.CurrentRow);
 
-                cmd.Parameters.AddWithValue("@Id", IdOff);
-                cmd.ExecuteNonQuery();
+                    //MENSAJE POR PANTALLA
+                    MessageBox.Show("Apagado automático removido!");
+                }
+                else
+                {
+
+                }
 
             }
 
-            dataGridViewOFF.Rows.Remove(dataGridViewOFF.CurrentRow);
-
-            //MENSAJE POR PANTALLA
-            MessageBox.Show("Apagado automático removido!");
+            
         }
 
         private void button1_Click_1(object sender, EventArgs e)
@@ -340,16 +367,25 @@ namespace SecadorBotas.Frames
                 {
                     MessageBox.Show("Apagado automático ya existe!");
                 }
-                else { 
+                else {
 
-                //LLAMADA DEL METODO INSERTAR APAGADO DE SECADOR 7
-                conex.InsertarApagado7(dia, hora, minuto);
+                    DialogResult result = MessageBox.Show("¿Esta Seguro Que Desea Guardar Esta programación?", "Confirmación", MessageBoxButtons.YesNo, MessageBoxIcon.Stop);
 
-                //LLAMADA A METODO QUE CARGA EL GRID
-                CargarDataGridOFF();
+                    if (result == DialogResult.Yes)
+                    {
+                        //LLAMADA DEL METODO INSERTAR APAGADO DE SECADOR 7
+                        conex.InsertarApagado7(dia, hora, minuto);
 
-                //MENSAJE POR PANTALLA
-                MessageBox.Show("Apagado automático guardado!");
+                        //LLAMADA A METODO QUE CARGA EL GRID
+                        CargarDataGridOFF();
+
+                        //MENSAJE POR PANTALLA
+                        MessageBox.Show("Apagado automático guardado!");
+                    }
+                    else
+                    {
+
+                    }
                 }
             }
         }
